Create Inorder Sucessor
Problem statement
You have been given an arbitrary binary tree and a node of this tree. You need to find the inorder successor of this node in the tree.

The inorder successor of a node in a binary tree is that node that will be visited immediately after the given node in the inorder traversal of the tree. If the given node is visited last in the inorder traversal, then its inorder successor is NULL.

The inorder traversal of a binary tree is the traversal method in which for any node its left subtree is visited first, then the node itself, and then the right subtree.

Note
1. Each node is associated with a unique integer value. 

2. The node for which the successor is to be found is guaranteed to be part of the tree.
Detailed explanation ( Input/output format, Notes, Images )
Constraints:
1 <= 'T' <= 100
2 <= 'N' <= 10 ^ 3  
1 <=  'NODEVALUE' <= 10 ^ 9

Where 'N' is the number of nodes and 'NODEVALUE' is the value of the node.

Time Limit: 1sec.
Sample Input 1:
1
1 2 3 4 5 -1 -1 -1 -1 -1 -1 
2
Sample Output 1:
5
Explanation of the Sample Input1:
The tree can be represented as follows:

The inorder traversal of this tree will be 4 2 5 1 3.
As the root is ‘1’, we first visit its left subtree which is rooted at 2, thus then we visit the left subtree of ‘2’, which is rooted at ‘4’. The left subtree of ‘4’ is NULL, so we visit the root i.e ‘4’. This process is continued.
So the order would be 4 -> 2 -> 5 -> 1 -> 3.
Clearly after the node with value = 2, the node with value = 5 is visited. So, 5 is the answer
Sample Input 2:
1
10 5 6 -1 -1 -1 -1
6
Sample Output 2:
NULL