Create Level Leaf
Problem statement
You are given a binary tree consisting of ‘N’ nodes. Ninja recently learned about tree algorithms and the teacher wants him to find whether all the leaves of the tree are at the same level.

A node is a leaf if it has neither the left child nor the right child.

If all the leaves are at the same level, output 1. Else, output 0.

Detailed explanation ( Input/output format, Notes, Images )
Constraints :
1 <= T <= 10
1 <= N <= 10^5
1 <= Value of node[i] <= 10^5
Sum of N over all Test cases <= 10^5
It is guaranteed that the given input is a binary tree.

Time Limit : 1 sec
Sample Input 1 :
2
3
1 2 3 -1 -1 -1 -1
4
1 2 3 4 -1 -1 -1 -1 -1  
Sample Output 1 :
1
0
Explanation Of Sample Input 1 :
For test case 1 we have, 

Nodes 2 and 3 are the leaves of the tree and they are at the same level.

So, we output 1.

For test case 2 we have,

The input tree is : 

Nodes 3 and 4 are the leaves of the tree at different levels.

So, we output 0.
Sample Input 2 :
2
5
2 2 -1 3 1 -1 4 -1 -1 -1 -1 
4
4 1 -1 1 1 -1 -1 -1 -1    
Sample Output 2 :
1
1