Create Count Messages
Problem statement
You are given a task to do some analysis for a messaging service application.

You are given series of queries of two types :

1 'user' 'time', denoting that the query is of type-1, and a person named 'user' has received a message at the time 'time'.

2 ‘l’ ‘r’ 'user' ‘k’, denoting that the query is of type-2, and [l, r] being a range of time, 'user' is a name of a person, ‘k’ is the chunk size.
For query of type-2 you need to print a list of integers, representing the number of messages the person 'user' received in each of the time chunks of size ‘k’, in time period [l, r].

Example:
If [l, r] = [10, 100] and ‘k’ = 35, then the chunks of size ‘k’ for the range will be:  [10, 44], [45, 79], [80, 100]. You need to return a list of size 3 here for each of these 3 chunks, where you tell the number of times a message is received by the person in that particular chunk (or sliced range).
Note :
Here you will implment function receiveMessage(user,time) for query of type-1, and implement function getMessageCount(l,r,user,k) for query fo type-2.
Detailed explanation ( Input/output format, Notes, Images )
Constraints :
1 <= q <= 100
1 <= |user| <= 10
1 <= k <= 100
0 <= l, r , time <= 3000

|user| denotes the length of the string ‘user’

Time Limit: 1 sec
Sample Input 1 :
5
1 abc 5
1 def 7
1 abc 9
1 abc 3
2 5 10 abc 3
Sample Output 1 :
1 1
Explanation For Sample Input 1 :
Before the query of type-2, we have 4 queries of type-1, where 3 are the ones where the user is “abc”, who received messages at the time [3, 5, 9].

We need to find messages in the range [5, 10] in chunks of 3. I.e. [5, 6, 7], [8, 9, 10].

The user “abc” received a message at timestamp 3: which doesn’t belong to any chunk

The user “abc” received a message at timestamp 5: which belongs to chunk 1

The user “abc” received a message at timestamp 7: which belongs to chunk 2.

Thus the answer is [1, 1], one message in each chunk.
Sample Input 2 :
6
1 nnax 85
1 nqwl 35
1 nnax 86
1 nqwl 30
1 nnax 1
2 5 43 nqwl 10
Sample Output 2 :
0 0 1 1