Create Magic Squares In Matrix
Problem statement
You are given a matrix ‘MAT’, of size ‘M’ * ‘N’ and has numbers from 1 to 9. Find the number of magic square submatrix present in the matrix. A submatrix is a magic square if that matrix's size is 3 * 3 and consists of distinct numbers from 1 to 9 and having the sum of all the rows, columns, and diagonals equal.

For example:
Let matrix be : 
8 3 4
1 5 9
6 7 2

The sum of the first row is: 15
The sum of the second row is: 15
The sum of the third row is: 15
The sum of the first column is: 15
The sum of the second column is: 15
The sum of the third column is: 15
The sum of both the diagonals is: 15
The matrix consists of distinct numbers from 1 to 9 and has the sum of all the rows, columns, and diagonals equal. So, it is a magic square.
Detailed explanation ( Input/output format, Notes, Images )
Constraints :
1 <= T <= 10
1 <= M, N <= 5*10^3
1 <= MAT[i][j] <= 9

Time Limit: 1 sec
Sample Input 1 :
2
3 3
8 3 4
1 5 9
6 7 2
3 3
1 3 5
1 6 3
7 9 2
Sample Output 1 :
1
0
Explanation For Sample Output 1 :
For test case 1: 
The sum of the first row is: 15
The sum of the second row is: 15
The sum of the third row is: 15
The sum of the first column is: 15
The sum of the second column is: 15
The sum of the third column is: 15
The sum of both the diagonals is: 15
The matrix consists of distinct numbers from 1 to 9 and has the sum of all the rows, columns, and diagonals equal. So, the count is 1.

For test case 2:    
The matrix does not consist of any magic squares.   
Sample Input 2 :
2
1 2
1 2
6 3
4 3 8
9 5 1
2 7 6
6 7 2
1 5 9
8 3 4
Sample Output 2 :
0
2