Create Binary Matrix
Problem statement
You are given a matrix ‘MAT’ consisting of ‘N’ rows and ‘M’ columns. Let (i, j) represent the cell at the intersection of the ith row and the jth column. Each cell of the matrix ‘MAT’ has either integer 0 or 1. For each cell in ‘MAT’, you have to find the Manhattan distance of the nearest cell from this cell that has the integer 0. The nearest cell will be the cell having the minimum Manhattan distance from it.

Manhattan distance between two cells, (p1, q1) and (p2, q2) is |p1 - p2| + |q1 - q2|.

You should return a matrix consisting of ‘N’ rows and ‘M’ columns, where the cell (i, j) represents the Manhattan distance of the nearest cell from the cell (i, j) in ‘MAT’ that has integer 0.

Note
1. There is at least one cell having the integer 0 in the given matrix.
Example:
Consider the following 2*3 matrix ‘MAT’:
                 [0, 1, 1]
                 [0, 1, 1]

Here, the nearest cell having the integer 0 from the cell (0, 0) is the cell (0, 0) itself. The Manhattan distance between them is |0 - 0| + |0 - 0| = 0.

The nearest cell having the integer 0 from the cell (0, 1) is cell (0, 0). The Manhattan distance between them is |0 - 0| + |1 - 0| = 1.

The nearest cell having the integer 0 from the cell (0, 2) is cell (0, 0). The Manhattan distance between them is |0 - 0| + |2 - 0| = 2.

The nearest cell having the integer 0 from the cell (1, 0) is cell (1, 0) itself. The Manhattan distance between them is |1 - 1| + |0 - 0| = 0.

The nearest cell having the integer 0 from the cell (1, 1) is cell (1, 0). The Manhattan distance between them is |1 - 1| + |1 - 0| = 1.

The nearest cell having the integer 0 from the cell (1, 2) is cell (1, 0). The Manhattan distance between them is |1 - 1| + |2 - 0| = 2.
Thus we should return matrix:

                [0, 1, 2]
                [0, 1, 2]
Detailed explanation ( Input/output format, Notes, Images )
Constraints:
1 <= T <= 50
1 <= N <= 100
1 <= M <= 100
0 <= MAT[i][j] <= 1


Where ‘T’ is the total number of test cases,  ‘N’ and ‘M’ denote the number of rows and columns in the given matrix ‘MAT’, and MAT[i][j] is the element of the given matrix at cell (i, j). 

Time limit: 1 sec.
Sample Input 1:
2
1 1
0
2 3
0 1 1
0 1 1
Sample Output 1:
0
0 1 2
0 1 2
Explanation of Sample Input 1:
Test case 1:
There is only one cell in the given matrix and that cell has integer 0, so the nearest cell from this cell having integer 0 is this cell itself. The Manhattan distance of the cell from itself will be 0.

Test case 2:
See the problem statement for an explanation.
Sample Input 2:
2
3 3
1 1 1
1 1 1
1 1 0
3 4
1 1 0 1
1 1 1 1
0 1 1 0
Sample Output 2:
4 3 2
2 3 1
2 1 0
2 1 0 1
1 2 1 1
0 1 1 0